ActiveDeck.tsx, CardCollection.tsx, EnergyCurve.tsx
