+++ src/common/components/cards/EnergyCurve.tsx
@@ -1,27 +1,31 @@
 import * as React from 'react';
-import { arrayOf, number, object } from 'prop-types';
 import { times } from 'lodash';
 
+import * as w from '../../types';
 import asyncComponent from '../AsyncComponent';
 
 const BarChart = asyncComponent(() => import(/* webpackChunkName: 'react-bar-chart' */ 'react-bar-chart'));
 
+interface EnergyCurveProps {
+  cards: w.CardInStore[]
+  height?: number
+}
+
+interface EnergyCurveState {
+  width: number
+}
+
 // Widget to display the current energy curve for a set of cards
-export default class EnergyCurve extends React.Component {
-  static propTypes = {
-    cards: arrayOf(object),
-    height: number
+export default class EnergyCurve extends React.Component<EnergyCurveProps, EnergyCurveState> {
+  public static defaultProps = {
+    height: 130
   };
 
-  static defaultProps = {
-    height: 130
-  }
-
-  state = {
+  public state = {
     width: 200
   };
 
-  componentDidMount() {
+  public componentDidMount(): void {
     this.updateWidth();
 
     window.addEventListener('resize', () => {
@@ -29,25 +33,18 @@ export default class EnergyCurve extends React.Component {
     });
   }
 
-  updateWidth() {
-    if (this.node) {
-      this.setState({
-        width: this.node.offsetWidth
-      });
-    }
-  }
+  get data(): Array<{ text: string, value: number }> {
+    const curve: Record<string, number> = {};
 
-  parseCards(cards) {
-    const curve = {};
-
-    cards.forEach(card => {
-      if (card.cost > 10)
+    this.props.cards.forEach((card) => {
+      if (card.cost > 10) {
         curve[10] ? curve[10] += 1 : curve[10] = 1;
-      else
+      } else {
         curve[card.cost] ? curve[card.cost] += 1 : curve[card.cost] = 1;
+      }
     });
 
-    const data = [];
+    const data: Array<{ text: string, value: number }> = [];
 
     times(10, (i) => {
       data.push({
@@ -64,7 +61,7 @@ export default class EnergyCurve extends React.Component {
     return data;
   }
 
-  render() {
+  public render(): JSX.Element {
     const margins = {
       top: 15,
       right: 10,
@@ -73,13 +70,21 @@ export default class EnergyCurve extends React.Component {
     };
 
     return (
-      <div ref={(node) => { this.node = node; }}>
+      <div ref={(node) => { (this as any).node = node; }}>
         <BarChart
           width={this.state.width}
           height={this.props.height}
           margin={margins}
-          data={this.parseCards(this.props.cards)} />
+          data={this.data} />
       </div>
     );
+  }
+
+  private updateWidth(): void {
+    if ((this as any).node) {
+      this.setState({
+        width: (this as any).node.offsetWidth
+      });
+    }
   }
 }
